+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The changelog for `Superwall`. Also see the [releases](https://github.com/superw
 ## Fixes
 
 - Fixes test mode products not being loaded properly in PW
+- Improve exponential backoff retry for Play Service unavailable
+- Fix Custom Info date serialization issues
 
 ## 2.7.13